requirements.txt
python-telegram-bot==20.6
python-telegram-bot[job-queue]==20.6
requests